# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Go modular monolith backend for emergency alert intake, ambulance dispatch, user management, and fleet readiness.",
        "title": "Dispatch Backend API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/incidents": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns paginated incidents",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Incidents"
                ],
                "summary": "List incidents",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Incident status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "District ID",
                        "name": "district_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Facility ID",
                        "name": "facility_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Priority level ID",
                        "name": "priority_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search by incident number, summary, or patient name",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "reported_at",
                            "created_at",
                            "status"
                        ],
                        "type": "string",
                        "description": "Sort field",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "ASC",
                            "DESC"
                        ],
                        "type": "string",
                        "description": "Sort order",
                        "name": "sort_order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates an incident and optionally persists triage responses on creation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Incidents"
                ],
                "summary": "Create incident with triage",
                "parameters": [
                    {
                        "description": "Incident payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dispatch_internal_modules_incidents_application.CreateIncidentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/incidents/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns an incident by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Incidents"
                ],
                "summary": "Get incident by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Incident ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/incidents/{id}/status": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates incident lifecycle status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Incidents"
                ],
                "summary": "Update incident status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Incident ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Incident status payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dispatch_internal_modules_incidents_application.UpdateIncidentStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dispatch_internal_modules_incidents_application.CreateIncidentRequest": {
            "type": "object",
            "required": [
                "incident_type_id",
                "source_channel",
                "summary"
            ],
            "properties": {
                "caller_name": {
                    "type": "string"
                },
                "caller_phone": {
                    "type": "string"
                },
                "created_by_user_id": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "district_id": {
                    "type": "string"
                },
                "facility_id": {
                    "type": "string"
                },
                "incident_type_id": {
                    "type": "string"
                },
                "landmark": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "parish": {
                    "type": "string"
                },
                "patient_age_group": {
                    "type": "string"
                },
                "patient_name": {
                    "type": "string"
                },
                "patient_phone": {
                    "type": "string"
                },
                "patient_sex": {
                    "type": "string"
                },
                "priority_level_id": {
                    "type": "string"
                },
                "questionnaire_code": {
                    "type": "string"
                },
                "severity_level_id": {
                    "type": "string"
                },
                "source_channel": {
                    "type": "string"
                },
                "subcounty": {
                    "type": "string"
                },
                "summary": {
                    "type": "string"
                },
                "triage_notes": {
                    "type": "string"
                },
                "triage_responses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dispatch_internal_modules_incidents_application.TriageResponseInput"
                    }
                },
                "village": {
                    "type": "string"
                }
            }
        },
        "dispatch_internal_modules_incidents_application.TriageResponseInput": {
            "type": "object",
            "required": [
                "question_code",
                "response_value"
            ],
            "properties": {
                "question_code": {
                    "type": "string"
                },
                "response_value": {
                    "type": "string"
                }
            }
        },
        "dispatch_internal_modules_incidents_application.UpdateIncidentStatusRequest": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "notes": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}