# internal/modules/incidents/application/dto.go
package application

import (
	"dispatch/internal/modules/incidents/domain"
	platformdb "dispatch/internal/platform/db"
)

type TriageResponseInput struct {
	QuestionCode  string `json:"question_code" binding:"required"`
	ResponseValue string `json:"response_value" binding:"required"`
}

type CreateIncidentRequest struct {
	SourceChannel     string                `json:"source_channel" binding:"required"`
	CallerName        string                `json:"caller_name"`
	CallerPhone       string                `json:"caller_phone"`
	PatientName       string                `json:"patient_name"`
	PatientPhone      string                `json:"patient_phone"`
	PatientAgeGroup   string                `json:"patient_age_group"`
	PatientSex        string                `json:"patient_sex"`
	IncidentTypeID    string                `json:"incident_type_id" binding:"required,uuid"`
	SeverityLevelID   *string               `json:"severity_level_id"`
	PriorityLevelID   *string               `json:"priority_level_id"`
	Summary           string                `json:"summary" binding:"required"`
	Description       string                `json:"description"`
	DistrictID        *string               `json:"district_id"`
	FacilityID        *string               `json:"facility_id"`
	Village           string                `json:"village"`
	Parish            string                `json:"parish"`
	Subcounty         string                `json:"subcounty"`
	Landmark          string                `json:"landmark"`
	Latitude          *float64              `json:"latitude"`
	Longitude         *float64              `json:"longitude"`
	CreatedByUserID   *string               `json:"created_by_user_id"`
	QuestionnaireCode string                `json:"questionnaire_code"`
	TriageResponses   []TriageResponseInput `json:"triage_responses"`
	TriageNotes       string                `json:"triage_notes"`
}

type UpdateIncidentStatusRequest struct {
	Status string `json:"status" binding:"required"`
	Notes  string `json:"notes"`
}

type ListIncidentsParams struct {
	Status     *string               `json:"status,omitempty"`
	DistrictID *string               `json:"district_id,omitempty"`
	FacilityID *string               `json:"facility_id,omitempty"`
	PriorityID *string               `json:"priority_id,omitempty"`
	Pagination platformdb.Pagination `json:"pagination"`
}

type CreateIncidentResponse struct {
	Incident                   domain.Incident                `json:"incident"`
	TriageSession              *domain.PersistedTriageSession `json:"triage_session,omitempty"`
	AutoDispatchEligible       bool                           `json:"auto_dispatch_eligible"`
	DispatchRecommendationHint string                         `json:"dispatch_recommendation_hint"`
}